fix error "updating spam/ham count: key too large" when marking some messages as ham/spam
when we train/untrain a message as ham or spam, we tokenize the message
into words, and increase/decrease the "seen in ham/spam message" count
for each word in the junk filter database. however, if a "word" is
larger than 32k, that operation would fail with "key too large", a
(reasonable) limitation of bbolt databases. we didn't check for the max
size of a word.

for issue #386